* add rotating and BobUpDown to marker

* Updated submodule runtime

* add metric

* Empty-Commit

* Update release.yml

* add Alt.Metric* calls

* add latest blip changes

* ILocalVehicle inherit ivehicle

* fix mvalue

* update to latest sdk changes

* fix VEG parameter name

* Added Create Ped api (#749)

Added Async Create Ped api


* mark all entity constructor as obsolete and add all Alt.Create*

* remove obsolete Alt.Server

* remove debug log

* add AltAsync.CreateVirtualEntity* and update natives

* fix byte entity type

* fix VE GetStreamSyncedMetaData

* fix StreamSyncedMetaChangeDelegate to baseobject

* ILocalPed extends IPed, all entities now have a remote id and isrmeote

* Updated submodule runtime

* add entity set pos/rot network check

* Disable current not supported types

* fix adapter check for mvalue

* Runtime update

* cleanup remoteid, cleanup scriptid

* optimize pipeline

* Runtime update

* Added missing event type

* prepare runtime with latest changes for weaponobject/font/frozen

* SDK update

* Fix localstorage null key crash

* update to latest sdk changes

* add latest events

* Add IPlayer::SetAmmo

* Update VehicleModelInfo and WeaponModelInfo

* Add mp2023_01 / San Andreas Mercenaries content (#753)

* add entity timestamp

* fix enum

* fix avenger4 (#754)

* Fix for weaponModelInfoCache not created

* Fix for weaponModelInfoCache not created (#755)


* Added AmmoType enum

* Adjusted names to be more like WeaponModel names

* update to latest runtime

* remove null check for netowner change

* Revert "remove null check for netowner change"

This reverts commit 5a463194679c330d4cb1436e9335155aa428b9f8.

* add set pos/dim for client

* Checkpoint type typo fix

* add missing parameter for create Object

* Add IsWebViewGpuAccelerationActive getter

* refactor blip getters/setters

* update priority

* add paraless constructor for position/rotation

* fix getcurrentweapon

* Updated submodule runtime

* add PLAYER_BULLET_HIT_EVENT event

* update runtime

* implement blip changes

* fix game crash by weapon damage

* Runtime update

* Runtime update

* update to latest SDK changes

* fix enum baseobjecttype

* add player Ammo methods

* add all audio classes to client

* Add SDK version to CI

* add serverside decorations

* Add Docker CI build

* fix OnRmlEventDelegate

* fix caller

* remove wrong event

* Fix clientside timers threadsafety

* add Add/Remove Output for Audio

* add AudioCategory

* add player playscenario

* Updated submodule runtime

* add all weapon data getter

* fix OnCheckpoint event

* Update runtime

* add missing client event register

* add CVoiceConnectionEvent and SetVoiceExternalPublic

* add SetVoiceExternal

* update to latest changes

* Change NetworkObject to Object

* add ResetMinimapComponentPosition

* update to latest change

* Add GetVoiceConnectionState

* Build fix

* add new blip apis

* fix check version

* fix check version

* fix var name Typo (#758)

rename streamingDistance to maxEntitiesInStream

* add cloud auth api

* Updated submodule runtime

* fix build

* Add some vehicle props

* Updated submodule runtime

* making all client factories virtual

* Runtime update

* Add GetAllResources to serverside

* Runtime update

* Fixed build

* runtime update

* EventType enum fix

* EventType enum fix

* add latest changes (3 scene events missing)

* fix event

* sync script events now need have the definded return type

* Runtime update

* add new events

* add FontRegister to client

* add OnceServer/OnceClient to Serverside

* add passengers to vehicle

* add connectioninfo Get/Set Text

* VoiceChannel_GetPlayerCount

* fix

* add IsParachuting

* OnPlayerWeaponChange is not cancable anymore

* Update runtime

* rc runtime

* Update runtime

* Update runtime

* fix runtime calls for audio

* add audio event handler

* add new core methods and add playerhealevent

* make WeaponDamageDelegate dynamic return

* small fixes

* Runtime update

* Runtime update

* Runtime update

* added missing SetDateTime method (#776)

* fixed GetBaseObjectById cast error (#777)

* added marker pool & factory (#778)

* fix valid type check

* add latest Core Method

* add Ped/Object for GetClosest

* Ped bone info fix

* add meta data setters for setting multiple values at once

* added object pool & renamed local object stuff (#779)

* added colshape create functions to client (#780)

* client: add SteeringAngle setter

* upgrade version bumps

* upgrade version bumps

* upgrade version bumps

* Updated submodule runtime

* Updated submodule runtime

* update runtime

* Updated submodule runtime

* make WeaponDamage Eventhandling better

* fix runtime and sample

* make it backwards comaptible

* fix checkpoint events and fix bool translate from runtime

* fix mvalue crash

* Extend get all functions (#785)

* runtime: added new get all functions

* runtime: added new free functions

* extend get all functions

* extend get all function in alt class

* updated librarys

* updated libraries

* updated libraries

* Updated submodule runtime

* added remaining get all functions to module wrapper (#801)

* Updated submodule runtime

* Updated submodule runtime

* Updated submodule runtime

* Update runtime

* ConnectionInfo.RequestCloudId

* Proper cloud auth impl

* Update runtime

* SDK Update

* revert a bit

* fix

* added give ped scripted task event (#803)

* update to latest sdk

* added virtual entity get all to alt class (#802)

* change CGivePedScriptedTaskEvent taskType to uint32_t

* Changed ScriptEventType.PlayerLeaveVehicle to ScriptEventType.PlayerEnteringVehicle

* Rename all Audio Outputs

* fix type

* Runtime update

* Runtime update

* add async entity factories. improve docs for both (#806)

* add latest events

* Updated submodule runtime

* add new eventtypes

* update runtime

* add RPC events

* fix event call

* update RPC to new answer type

* fix mock

* Exposed RegisterFont function

* fix SocialId name in connectionInfo

* Update voice settings image

* Updated submodule runtime

* make rpc events shared

* streaming distance per entity

* Runtime update

* Implemented clearCache argument in Audio creation

* add entity SetStreamingDistance

* add BaseObjectTypeSize check

* refactor cloud auth

* Added GetLoadedVehicleModels

* fix mock

* move EmitRpc to IPlayer

* move EmitRpc to IPlayer

* add missing async events

* by parsing null string, translate it to empty one

* only use MemoryUtils

* Revert "by parsing null string, translate it to empty one"

This reverts commit 45b6ee0fdbf4729732553decbd6bea6fbb0bc71d.

* Auto stash before revert of "by parsing null string, translate it to empty one"

* Add NetTime as Alt.NetTime for client and server

* Added method to pass more config entries to client

* add cloud auth result and benefit check

* add blood damage getter / setter

* fix weaponDamage by damage value

* fix GetEntitiesIn*

* fix getid data type

* update runtime

---------

+++ api/AltV.Net.Client/Elements/Factories/PlayerFactory.cs
@@ -5,12 +5,12 @@ namespace AltV.Net.Client.Elements.Factories
 {
     public class PlayerFactory : IPlayerFactory
     {
-        public IPlayer Create(ICore core, IntPtr playerPointer, ushort id)
+        public IPlayer Create(ICore core, IntPtr playerPointer, uint id)
         {
             return new Player(core, playerPointer, id);
         }
 
-        public ILocalPlayer GetLocalPlayer(ICore core, IntPtr localPlayerPointer, ushort id)
+        public ILocalPlayer GetLocalPlayer(ICore core, IntPtr localPlayerPointer, uint id)
         {
             return new LocalPlayer(core, localPlayerPointer, id);
         }
