+++ api/AltV.Net.Mock/MockPlayer.cs
@@ -5,7 +5,7 @@ namespace AltV.Net.Mock
 {
     public class MockPlayer : Player
     {
-        public MockPlayer(ICore core, IntPtr nativePointer, ushort id) : base(core, nativePointer, id)
+        public MockPlayer(ICore core, IntPtr nativePointer, uint id) : base(core, nativePointer, id)
         {
         }
     }
