+++ api/AltV.Net.Mock/MockVehicle.cs
@@ -3,9 +3,9 @@ using AltV.Net.Elements.Entities;
 
 namespace AltV.Net.Mock
 {
-    public class MockVehicle: Vehicle
-    {   
-        public MockVehicle(ICore core, IntPtr nativePointer, ushort id): base(core, nativePointer, id)
+    public class MockVehicle : Vehicle
+    {
+        public MockVehicle(ICore core, IntPtr nativePointer, uint id) : base(core, nativePointer, id)
         {
         }
     }
